ejercicios de hilos en SO

+++ sistemas operativos/señales/practica1/ejercicio8/envia.c
@@ -7,14 +7,17 @@
 #include <string.h>
 #include <signal.h>
 
-int main(int argc, char ** argv){
+int main(int argc, char **argv)
+{
+    if (argc != 2)
+    {
+        printf("Debe introducir el pid de un proceso obligatoriamente.\n");
+        exit(EXIT_FAILURE);
+    }
     int pid = atoi(argv[1]);
     printf("Enviando señal SIGUSR1 a %d ...\n", pid);
-    //envia una señal a la que tiene tal pid
     kill(pid, SIGUSR1);
-    //duerme un segundo
     sleep(1);
-    //mata el proceso
     kill(pid, SIGKILL);
     exit(EXIT_SUCCESS);
-}   
+}
